+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 
-## v1.0.25 ~
+## v1.0.26 ~ 5-3-21
+
+  * New Module: DivSeq
+  * 8Seq, GridSeq, DivSeq: can now Alt + click randomize notes, divs, probability - this will set the max to the top left knob corresponding to the parameter being randomized
+
+## v1.0.25 ~ 4-6-21
 
   * 8seq length issue fixed